Rename MySQL-Support to MariaDB-Support
