+++ mariadb/src/test/scala/org/scalarelational/mariadb/MariaDBTest.scala
@@ -1,23 +1,23 @@
-package org.scalarelational.mysql
+package org.scalarelational.mariadb
 
-import org.scalarelational.mysql.MySQLTestDataStore$._
-import org.scalarelational.mysql.MySQLTestDataStore$.suppliers._
-import org.scalarelational.mysql.MySQLTestDataStore$.{coffees, suppliers}
+import org.scalarelational.mariadb.MariaDBTestDataStore$._
+import org.scalarelational.mariadb.MariaDBTestDataStore$.suppliers._
+import org.scalarelational.mariadb.MariaDBTestDataStore$.{coffees, suppliers}
 
 // TODO: remove this as soon as the TableSpec is working
 @Deprecated
-object MysqlTest extends App {
+object MariaDBTest extends App {
 
   override def main(args: Array[String]) {
     session {
       create(suppliers, coffees)
 
       //Create suppliers
-      val acmeId = MySQLTestDataStore$.insert(name("Acme, Inc."), street("99 Market Street"), city("Groundsville"), state("CA"), zip("95199")).result
+      val acmeId = MariaDBTestDataStore$.insert(name("Acme, Inc."), street("99 Market Street"), city("Groundsville"), state("CA"), zip("95199")).result
       val superiorCoffeeId = insert(name("Superior Coffee"), street("1 Party Place"), city("Mendocino"), state("CA"), zip("95460")).result
       val theHighGroundId = insertInto(suppliers, "The High Ground", "100 Coffee Lane", "Meadows", "CA", "93966").result
 
-      import org.scalarelational.mysql.MySQLTestDataStore$.coffees._
+      import org.scalarelational.mariadb.MariaDBTestDataStore$.coffees._
 
       session {
         // Batch insert some coffees
